Ajout de l'authorization aux routes

+++ backend/src/resolvers/ad.resolver.ts
@@ -20,26 +20,41 @@ export class AdResolver {
         console.log(ctx)
         return AdService.search(categoryId, search);
     }
-    
+
     @Query(() => Ad)
-    getAd(@Arg("id") id: number): Promise<Ad | null>{
+    getAd(@Arg("id") id: number): Promise<Ad | null> {
         return AdService.findAdById(id);
     }
 
+    @Authorized()
     @Mutation(() => String)
-    async deleteAd(@Arg("id") id: number): Promise<string>{
-        await AdService.deleteAd(id);        
+    async deleteAd(
+        @Arg("id") id: number,
+        @Ctx() ctx: Context | any
+    ): Promise<string> {
+        const ad = await AdService.findAdById(id);        
+        if(ad?.owner !== ctx.user.email && ctx.user.roles !== "ADMIN"){
+            return "Not authorized"
+        } 
+        
+        await AdService.deleteAd(id);
         return "Annonce supprimée";
     }
 
     @Mutation(() => Ad)
-    createAd(@Arg("ad") ad: CreateAdInputType): Promise<Ad>{
-        return AdService.create({...ad});
+    @Authorized()
+    createAd(
+        @Arg("ad") ad: CreateAdInputType,
+        @Ctx() ctx: Context
+    ): Promise<Ad> {
+        return AdService.create({ ...ad }, ctx);
     }
 
+    @Authorized()
     @Mutation(() => Ad)
-    updateAd(@Arg("ad") ad: UpdateAdInputType, @Arg("categoryId") categoryId: number): Promise<Ad | undefined> {
-            return AdService.update(ad.id, {...ad} as Ad, categoryId)
-        }
-    
+    updateAd(@Arg("ad") ad: UpdateAdInputType, @Arg("categoryId") categoryId: number, @Ctx() ctx: Context | any): Promise<Ad | undefined> {
+
+        return AdService.update(ad.id, { ...ad } as Ad, categoryId, ctx)
+    }
+
 }
